requirements.txt
anthropic
feedparser
pyyaml
markdown
flask
python-dotenv
openai
scipy
pystray
Pillow
